+++ test/DebugInfo/inlinescopes.swift
@@ -1,9 +1,9 @@
 // RUN: rm -rf %t
 // RUN: mkdir %t
-// RUN: echo "public var x = Int()" | %swift -target x86_64-apple-macosx10.9 -module-name FooBar -emit-module -o %t -
-// RUN: %swift -target x86_64-apple-macosx10.9 %s -O -I=%t -sil-inline-threshold 100 -emit-ir -g -o %t.ll
-// RUN: cat %t.ll | FileCheck %s
-// RUN: cat %t.ll | FileCheck %s -check-prefix=TRANSPARENT-CHECK
+// RUN: echo "public var x = Int64()" | %target-swift-frontend -module-name FooBar -emit-module -o %t -
+// RUN: %target-swift-frontend %s -O -I %t -sil-inline-threshold 100 -emit-ir -g -o %t.ll
+// RUN: FileCheck %s < %t.ll
+// RUN: FileCheck %s -check-prefix=TRANSPARENT-CHECK < %t.ll
 
 // CHECK: define i32 @main
 // CHECK: tail call { i64, i1 } @llvm.smul.with.overflow.i64(i64 %[[C:.*]], i64 %[[C]]), !dbg ![[MULSCOPE:.*]]
@@ -13,7 +13,7 @@
 
 import FooBar
 
-func square(x : Int) -> Int {
+func square(x: Int64) -> Int64 {
 // CHECK-DAG: ![[MULSCOPE]] = !MDLocation(line: [[@LINE+2]], column: {{.*}}, scope: ![[MUL:.*]], inlinedAt: ![[INLINED:.*]])
 // CHECK-DAG: ![[MUL:.*]] = {{.*}}[ DW_TAG_lexical_block ]
   let res = x * x
@@ -21,11 +21,11 @@ func square(x : Int) -> Int {
 // TRANSPARENT-CHECK-NOT: [ DW_TAG_subprogram ] {{.*}}[_TFSsoi1mFTSiSi_Si]
   return res
 }
-let c = Int(x)
-// CHECK-DAG ![[INLINED]] = !MDLocation(line: [[@LINE+1]], column: {{.*}}, scope: !{{.*}}, inlinedAt: ![[INLINED_TOPLEVEL:.*]])
+let c = Int64(x)
+// CHECK-DAG ![[INLINED]] = !MDLocation(i32 [[@LINE+1]], column: {{.*}}, scope: !{{.*}}, inlinedAt: ![[INLINED_TOPLEVEL:.*]])
 // CHECK-DAG: ![[TOPLEVEL]]{{.*}} ; [ DW_TAG_variable ] [y] [line [[@LINE+1]]]
 let y = square(c)
 println(y)
 
 // Check if the inlined and removed square function still has the correct linkage name in the debug info.
-// CHECK-DAG: _TF4main6squareFSiSi{{.*}}; [ DW_TAG_subprogram ] {{.*}}[square]
+// CHECK-DAG: _TF4main6squareFVSs5Int64S0_{{.*}}; [ DW_TAG_subprogram ] {{.*}}[square]
