+++ test/DebugInfo/shadow_copies.swift
@@ -1,10 +1,11 @@
-// RUN: %swift -target x86_64-apple-macosx10.9 %s -Onone -emit-ir -g -o - | FileCheck %s
+// RUN: %target-swift-frontend %s -Onone -emit-ir -g -o - | FileCheck %s
+
 class ClassA
 {
-    var x : Int
+    var x : Int64
     var y : Float
 
-    init (_ input : Int)
+    init (_ input : Int64)
     {
         x = input
         y = Float(input) + 0.5
@@ -13,9 +14,9 @@ class ClassA
 
 class ClassB : ClassA
 {
-    override init (_ input : Int)
+    override init (_ input : Int64)
     {
-    // CHECK: @_TFC{{.*}}6ClassBcfMS0_FSiS0_
+    // CHECK: @_TFC{{.*}}6ClassBcfMS0_FVSs5Int64S0_
     // CHECK:  alloca {{.*}}ClassB*
     // CHECK:  alloca i64
     // CHECK-NOT: alloca
