+++ test/IRGen/clang_inline.swift
@@ -1,11 +1,13 @@
 // RUN: rm -rf %t && mkdir -p %t
-// RUN: %swift -target x86_64-apple-macosx10.9 -sdk %S/Inputs -primary-file %s -emit-ir | FileCheck %s
+// RUN: %target-swift-frontend -sdk %S/Inputs -primary-file %s -emit-ir | FileCheck %s
 
 // RUN: mkdir -p %t/Empty.framework/Modules/Empty.swiftmodule
-// RUN: %swift -target x86_64-apple-macosx10.9 -emit-module-path %t/Empty.framework/Modules/Empty.swiftmodule/x86_64.swiftmodule %S/../Inputs/empty.swift -module-name Empty
-// RUN: %swift -target x86_64-apple-macosx10.9 -sdk %S/Inputs -primary-file %s -F %t -DIMPORT_EMPTY -emit-ir > %t.ll
+// RUN: %target-swift-frontend -emit-module-path %t/Empty.framework/Modules/Empty.swiftmodule/%target-swiftmodule-name %S/../Inputs/empty.swift -module-name Empty
+// RUN: %target-swift-frontend -sdk %S/Inputs -primary-file %s -F %t -DIMPORT_EMPTY -emit-ir > %t.ll
 // RUN: FileCheck %s < %t.ll
 // RUN: FileCheck -check-prefix=NEGATIVE %s < %t.ll
+
+// REQUIRES: CPU=i386_or_x86_64
 
 #if IMPORT_EMPTY
 import Empty
@@ -13,16 +15,16 @@ import Empty
 
 import gizmo
 
-// CHECK: define hidden i64 @_TFC12clang_inline16CallStaticInline10ReturnZerofS0_FT_Si(%C12clang_inline16CallStaticInline*) {
+// CHECK: define hidden i64 @_TFC12clang_inline16CallStaticInline10ReturnZerofS0_FT_VSs5Int64(%C12clang_inline16CallStaticInline*) {
 // CHECK: define internal i32 @zero() #0 {
 class CallStaticInline {
-  func ReturnZero() -> Int { return Int(zero()) }
+  func ReturnZero() -> Int64 { return Int64(zero()) }
 }
 
-// CHECK: define hidden i64 @_TFC12clang_inline17CallStaticInline210ReturnZerofS0_FT_Si(%C12clang_inline17CallStaticInline2*) {
+// CHECK: define hidden i64 @_TFC12clang_inline17CallStaticInline210ReturnZerofS0_FT_VSs5Int64(%C12clang_inline17CallStaticInline2*) {
 // CHECK: define internal i32 @wrappedZero() #0 {
 class CallStaticInline2 {
-  func ReturnZero() -> Int { return Int(wrappedZero()) }
+  func ReturnZero() -> Int64 { return Int64(wrappedZero()) }
 }
 
 // CHECK: define hidden i32 @_TF12clang_inline10testExternFT_VSs5Int32() {
